ad_op: change addr_t -> int and fix conversion warnings.

+++ ad_op/type.hpp
@@ -9,7 +9,7 @@
 # include <cppad/utility/vector.hpp>
 
 // addr_type
-typedef size_t addr_t;
+typedef int addr_t;
 
 // Vector
 template <class Base> using Vector = CppAD::vector<Base>;
